pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "questrade-price-fetcher"
version = "1.0.0"
requires-python = ">=3.11"
description = "Fetches real-time security prices from the Questrade API"
dependencies = ["httpx>=0.27.0", "pydantic>=2.7.0", "ttkbootstrap>=1.10.0", "pystray>=0.19.0", "Pillow>=10.0.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["*/__init__.py"]

[tool.coverage.report]
fail_under = 80

[tool.mypy]
strict = true
python_version = "3.11"
mypy_path = "src"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "ANN", "B", "SIM"]
ignore = ["ANN101", "ANN102"]
